+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [0ver](https://0ver.org).
 
 ## [Unreleased]
 
+### Added
+
+- Added automatic refresh of available projects when using wildcards
+- Added support for wildcard on refs (branches & tags) with automatic refresh of available ones
+
 ## [0.2.1] - 2019-07-26
 
 ### Added